app/requirements.txt
fastapi
torch
torchvision
uvicorn
pillow

